backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-groq==0.0.1
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.2

# RAG Dependencies
pypdf==3.17.4
faiss-cpu==1.7.4
sentence-transformers>=2.6.0
